// src/interfaces/product.ts
// Generated by https://quicktype.io

export interface ProductListResponse {
  RESPONSE: Response;
}

export interface Response {
  MENSAJE: Mensaje;
  PRODUCTOS: Productos;
}

export interface Mensaje {
  CODIGO: Codigo;
  TEXTO: Codigo;
}

export interface Codigo {
  text: string;
}

export interface Productos {
  producto: Producto[];
}

export interface Producto {
  attributes: {
    servicio: string;
    producto: string;
    idServicio: string;
    idProducto: string;
    idCatTipoServicio: string;
    tipoFront: TipoFront;
    hasDigitoVerificador: string;
    precio: string;
    showAyuda: "false" | "true";
    tipoReferencia: TipoReferencia;
  };
  legend: Legend;
}

export type TipoFront = "1" | "2" | "4" | "5";

export interface Legend {
  cdata: string;
}

export enum TipoReferencia {
  A = "a",
  B = "b",
  Bc = "bc",
  Ab = "ab",
  C = "c",
}